Refactor aggregation out of visualizer_system
