+++ crates/re_space_view_time_series/src/space_view_class.rs
@@ -15,7 +15,8 @@ use re_viewer_context::{
     SystemExecutionOutput, ViewQuery, ViewerContext,
 };
 
-use crate::visualizer_system::{PlotSeriesKind, TimeSeriesSystem};
+use crate::visualizer_system::TimeSeriesSystem;
+use crate::PlotSeriesKind;
 
 // ---
